+++ src/components/UI/FundingPage/FundingInfo.tsx
@@ -1,8 +1,9 @@
-import React, { JSX } from 'react'
+import React, { JSX, useState } from 'react'
 import ExamImage from '../../../assets/images/nextlevel.png'
 import LikeImage from '../../../assets/images/Like.svg'
 import { api } from '../../../AxiosInstance'
 import { useParams } from 'react-router-dom'
+import { FundingOptionManageModal } from './modals/FundingOptionManageModal'
 
 interface props {
 	setPayOpen: React.Dispatch<React.SetStateAction<boolean>>
@@ -14,14 +15,20 @@ interface props {
 	peopleNum: number
 	likeNum: number
 	tag: []
+	isAuthor: boolean
 }
 
-const FundingInfo = ({ setPayOpen, title, percent, image, description, amount, peopleNum, likeNum, tag }: props): JSX.Element => {
+const FundingInfo = ({ setPayOpen, title, percent, image, description, amount, peopleNum, likeNum, tag, isAuthor }: props): JSX.Element => {
 	const baseUrl = process.env.REACT_APP_API_BASE_URL
 	const { no } = useParams<{ no: string }>()
+	const [isOptionModalOpen, setIsOptionModalOpen] = useState(false)
 
 	const PayClick = () => {
 		setPayOpen(true)
+	}
+
+	const handleOptionManage = () => {
+		setIsOptionModalOpen(true)
 	}
 
 	const handleLike = async () => {
@@ -63,12 +70,24 @@ const FundingInfo = ({ setPayOpen, title, percent, image, description, amount, p
 					<img src={LikeImage} className='w-12 h-12 mb-1.5 hover:cursor-pointer transition-transform duration-200 group-hover:scale-110' onClick={handleLike} alt='좋아요' />
 					<p className='flex justify-center text-sm m-0 text-gray-500'>{likeNum}</p>
 				</div>
-				<button
-					className='bg-purple-500 hover:bg-purple-600 text-white border-none rounded-xl text-xl font-bold w-full h-12 hover:cursor-pointer transition-all duration-200 hover:shadow-lg hover:shadow-purple-500/30 hover:-translate-y-0.5'
-					onClick={PayClick}>
-					스타터와 함께하기
-				</button>
+				<div className='flex gap-2.5 ml-auto w-full'>
+					<button
+						className='bg-purple-500 hover:bg-purple-600 text-white border-none rounded-xl text-sm font-bold w-full h-12 hover:cursor-pointer transition-all duration-200 hover:shadow-lg hover:shadow-purple-500/30 hover:-translate-y-0.5'
+						onClick={PayClick}>
+						스타터와 함께하기
+					</button>
+					{isAuthor && (
+						<button
+							className='border-purple-500 text-purple-500 border-2 rounded-xl text-sm font-bold w-1/3 h-12 hover:cursor-pointer transition-all duration-200 hover:shadow-lg hover:shadow-purple-500/30 hover:-translate-y-0.5'
+							onClick={handleOptionManage}>
+							...
+						</button>
+					)}
+				</div>
 			</div>
+
+			{/* 옵션 관리 모달 */}
+			<FundingOptionManageModal isOpen={isOptionModalOpen} onClose={() => setIsOptionModalOpen(false)} projectId={no || ''} />
 		</div>
 	)
 }
