+++ src/pages/FundingPage.tsx
@@ -1,17 +1,24 @@
-import React, { JSX, useState } from 'react'
+import React, { JSX, useEffect, useState } from 'react'
 import FundingInfo from '../components/UI/FundingPage/FundingInfo'
 import StarterInfo from '../components/UI/FundingPage/StarterInfo'
 import FundingContent from '../components/UI/FundingPage/FundingContent'
 import FundingModal from '../components/UI/FundingPage/modals/FundingModal'
 import Modal from '../components/layout/Modal'
 import { useParams } from 'react-router-dom'
-import { useProjectDetailFetch, useProjectFetch } from '../apis/useProjectFetch'
+import { useProjectDetailFetch, useProjectFetch } from '../apis/funding/useProjectFetch'
+import { useAuth } from '../hooks/AuthContext'
 
 const FundingPage = (): JSX.Element => {
 	const { no } = useParams<{ no: string }>()
 	const [payOpen, setPayOpen] = useState<boolean>(false)
 	const { story, notice, community } = useProjectFetch({ projectId: no ?? '' })
 	const { projectInfo } = useProjectDetailFetch({ projectId: no ?? '' })
+	const { user } = useAuth()
+	const [isAuthor, setIsAuthor] = useState<boolean>(false)
+
+	useEffect(() => {
+		setIsAuthor(user?.nickName === projectInfo?.user?.nickName)
+	}, [projectInfo])
 
 	return (
 		<div className='flex pl-[2%] gap-[2%] mx-[15%] mt-10'>
@@ -26,6 +33,7 @@ const FundingPage = (): JSX.Element => {
 					peopleNum={projectInfo?.fundingCount ?? 0}
 					likeNum={projectInfo?.likeCount ?? 0}
 					tag={projectInfo?.tag ?? []}
+					isAuthor={isAuthor}
 				/>
 				<StarterInfo starter={projectInfo?.user?.nickName ?? ''} />
 			</div>
